Refactor leftover vrang classes
