+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/PackageStoreFactory.java
@@ -23,14 +23,18 @@ import org.slf4j.LoggerFactory;
 
 import com.vmware.pscoe.iac.artifact.aria.automation.configuration.ConfigurationAbx;
 import com.vmware.pscoe.iac.artifact.aria.automation.configuration.ConfigurationVraNg;
-import com.vmware.pscoe.iac.artifact.aria.automation.pack.VraNgPackageStore;
 import com.vmware.pscoe.iac.artifact.aria.automation.rest.RestClientVraNg;
 import com.vmware.pscoe.iac.artifact.aria.automation.store.AbxPackageStore;
+import com.vmware.pscoe.iac.artifact.aria.automation.store.VraNgPackageStore;
+import com.vmware.pscoe.iac.artifact.aria.codestream.configuration.ConfigurationCs;
+import com.vmware.pscoe.iac.artifact.aria.codestream.rest.RestClientCs;
+import com.vmware.pscoe.iac.artifact.aria.codestream.store.CsPackageStore;
 import com.vmware.pscoe.iac.artifact.aria.logs.configuration.ConfigurationVrli;
 import com.vmware.pscoe.iac.artifact.aria.logs.rest.v1.RestClientVrliV1;
 import com.vmware.pscoe.iac.artifact.aria.logs.rest.v2.RestClientVrliV2;
 import com.vmware.pscoe.iac.artifact.aria.logs.store.v1.VrliPackageStoreV1;
 import com.vmware.pscoe.iac.artifact.aria.logs.store.v2.VrliPackageStoreV2;
+import com.vmware.pscoe.iac.artifact.aria.operations.cli.CliManagerVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.configuration.ConfigurationVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.rest.RestClientVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.store.VropsPackageStore;
@@ -38,13 +42,11 @@ import com.vmware.pscoe.iac.artifact.aria.orchestrator.configuration.Configurati
 import com.vmware.pscoe.iac.artifact.aria.orchestrator.configuration.ConfigurationVroNg;
 import com.vmware.pscoe.iac.artifact.aria.orchestrator.rest.RestClientVro;
 import com.vmware.pscoe.iac.artifact.aria.orchestrator.store.VroPackageStore;
-import com.vmware.pscoe.iac.artifact.cli.CliManagerFactory;
-import com.vmware.pscoe.iac.artifact.cli.CliManagerVrops;
+import com.vmware.pscoe.iac.artifact.bsc.configuration.ConfigurationSsh;
+import com.vmware.pscoe.iac.artifact.bsc.store.SshPackageStore;
+import com.vmware.pscoe.iac.artifact.common.cli.CliManagerFactory;
 import com.vmware.pscoe.iac.artifact.configuration.Configuration;
-import com.vmware.pscoe.iac.artifact.configuration.ConfigurationCs;
-import com.vmware.pscoe.iac.artifact.configuration.ConfigurationSsh;
 import com.vmware.pscoe.iac.artifact.model.Version;
-import com.vmware.pscoe.iac.artifact.rest.RestClientCs;
 import com.vmware.pscoe.iac.artifact.rest.RestClientFactory;
 import com.vmware.pscoe.iac.artifact.strategy.Strategy;
 import com.vmware.pscoe.iac.artifact.strategy.StrategyForceLatestVersions;
