+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/aria/automation/store/VraNgPackageStore.java
@@ -16,7 +16,7 @@
  * LICENSE file.
  * #L%
  */
-package com.vmware.pscoe.iac.artifact.aria.automation.pack;
+package com.vmware.pscoe.iac.artifact.aria.automation.store;
 
 import java.io.File;
 import java.io.IOException;
@@ -25,20 +25,19 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.vmware.pscoe.iac.artifact.GenericPackageStore;
 import com.vmware.pscoe.iac.artifact.PackageManager;
 import com.vmware.pscoe.iac.artifact.PackageStore;
 import com.vmware.pscoe.iac.artifact.aria.automation.configuration.ConfigurationVraNg;
-import com.vmware.pscoe.iac.artifact.model.Package;
-import com.vmware.pscoe.iac.artifact.model.PackageContent.Content;
+import com.vmware.pscoe.iac.artifact.aria.automation.rest.RestClientVraNg;
 import com.vmware.pscoe.iac.artifact.aria.automation.store.models.VraNgPackageAdapter;
 import com.vmware.pscoe.iac.artifact.aria.automation.store.models.VraNgPackageContent;
 import com.vmware.pscoe.iac.artifact.aria.automation.store.models.VraNgPackageDescriptor;
-import com.vmware.pscoe.iac.artifact.aria.automation.rest.RestClientVraNg;
-import com.vmware.pscoe.iac.artifact.aria.automation.store.VraNgTypeStoreFactory;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.vmware.pscoe.iac.artifact.model.Package;
+import com.vmware.pscoe.iac.artifact.model.PackageContent.Content;
 
 public class VraNgPackageStore extends GenericPackageStore<VraNgPackageDescriptor> {
 	/**
